+++ src/components/Meal/MealItem/MealItem.js
@@ -1,4 +1,5 @@
 import styles from "./MealItem.module.css";
+import MealItemForm from "./MealItemForm";
 
 const MealItem = (props) => {
 	const price = `$${props.price.toFixed(2)}`;
@@ -10,7 +11,7 @@ const MealItem = (props) => {
 				<div className={styles["meal-description"]}>{props.description}</div>
 				<div className={styles["meal-price"]}>{price}</div>
 			</div>
-			<div></div>
+			<MealItemForm />
 		</li>
 	);
 };